Rename packages of cmp modules
