+++ lbccore-cmp/src/commonMain/kotlin/studio/lunabee/compose/core/cmp/LbcTextSpec.kt
@@ -16,10 +16,10 @@
  * LbcTextSpec.kt
  * Lunabee Compose
  *
- * Created by Lunabee Studio / Date - 11/19/2024 - for the Lunabee Compose library.
+ * Created by Lunabee Studio / Date - 11/20/2024 - for the Lunabee Compose library.
  */
 
-package studio.lunabee.compose.core.kmp
+package studio.lunabee.compose.core.cmp
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
@@ -28,7 +28,7 @@ import org.jetbrains.compose.resources.ExperimentalResourceApi
 import org.jetbrains.compose.resources.PluralStringResource
 import org.jetbrains.compose.resources.pluralStringResource
 import org.jetbrains.compose.resources.stringResource
-import studio.lunabee.compose.core.kmp.ext.format
+import studio.lunabee.compose.core.cmp.ext.format
 import kotlin.collections.contentEquals
 import kotlin.collections.contentHashCode
 import kotlin.collections.isEmpty
