+++ lbccore-cmp/src/commonMain/kotlin/studio/lunabee/compose/core/cmp/ext/StringExt.kt
@@ -19,7 +19,7 @@
  * Created by Lunabee Studio / Date - 11/19/2024 - for the Lunabee Compose library.
  */
 
-package studio.lunabee.compose.core.kmp.ext
+package studio.lunabee.compose.core.cmp.ext
 
 /**
  * Format a string with given parameters.
